# server/generated/openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "prompton-api",
    "description": "API for prompton - managing full lifecycle of AI chat prompts.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Users List",
        "operationId": "get_users_list",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Get Users List Users Get",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserRead"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Add New User",
        "operationId": "add_new_user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Item created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultPostResponse"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          },
          "409": {
            "description": "Email already exists"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get Current User",
        "operationId": "get_current_user",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User By Id",
        "operationId": "get_user_by_id",
        "parameters": [
          {
            "required": true,
            "schema": {
              "title": "Id",
              "type": "string"
            },
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            }
          },
          "422": {
            "description": "Invalid itemid supplied"
          },
          "404": {
            "description": "Item not found"
          },
          "400": {
            "description": "Malformed request"
          },
          "401": {
            "description": "Not authenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DefaultPostResponse": {
        "title": "DefaultPostResponse",
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string",
            "description": "The id of the created item"
          }
        },
        "description": "Standard response for POST requests. Contains the id of the created item."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "UserCreate": {
        "title": "UserCreate",
        "required": [
          "org_id",
          "email",
          "plain_password"
        ],
        "type": "object",
        "properties": {
          "full_name": {
            "title": "Full Name",
            "minLength": 1,
            "type": "string"
          },
          "disabled": {
            "title": "Disabled",
            "type": "boolean",
            "default": false
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoles"
              }
            ],
            "default": "Basic"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string",
            "format": "email"
          },
          "plain_password": {
            "title": "Plain Password",
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UserRead": {
        "title": "UserRead",
        "required": [
          "_id",
          "created_at",
          "created_by_user_id",
          "created_by_org_id",
          "disabled",
          "org_id",
          "email"
        ],
        "type": "object",
        "properties": {
          "_id": {
            "title": " Id",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string",
            "format": "date-time"
          },
          "created_by_user_id": {
            "title": "Created By User Id",
            "type": "string"
          },
          "created_by_org_id": {
            "title": "Created By Org Id",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "minLength": 1,
            "type": "string"
          },
          "disabled": {
            "title": "Disabled",
            "type": "boolean"
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRoles"
              }
            ],
            "default": "Basic"
          },
          "org_id": {
            "title": "Org Id",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          }
        },
        "description": "Base model for reading from MongoDB. Same as MongoBaseCreate but assumes all DB base fields are populated so generated clients doesn't requrie None checks"
      },
      "UserRoles": {
        "title": "UserRoles",
        "enum": [
          "Basic",
          "PromptAdmin",
          "OrgAdmin",
          "SuperAdmin"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  }
}